scripts/python/requirements.txt
langchain==0.3.13
langchain-google-genai==2.1.1
python-dotenv==1.0.0
tqdm==4.67.1
langchain-openai==0.2.11
